Introduce buffered file watcher and update error reporting
Block errors will now appear in the block itself during iteration.
Critical program errors will still end the program. This change paves
the way for removing `name` from the parse error (which remains to be
done, it should be replaced by `Parse2`).

A new watcher (`watch2`) has been introduced which holds a file pointer
and reuses a buffer on the stack to read to. Reducing the number of
memory allocations and system file opens. There is still work to be done
to replace `read_to_ty` usages with this function.